+++ test/StreamCoreAdminCharacterization.t.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.19;
 
 import "../smart-contracts/DependencyRegistry.sol";
 import "../smart-contracts/StreamAdmins.sol";
-import "../smart-contracts/StreamCore.sol";
+import {LegacyStreamCore as StreamCore} from "./helpers/LegacyStreamCore.sol";
 import "./helpers/Assertions.sol";
 import "./helpers/CharacterizationTestBase.sol";