+++ test/StreamDropsIntegrationCharacterization.t.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.19;
 
-import "../smart-contracts/StreamCore.sol";
+import {LegacyStreamCore as StreamCore} from "./helpers/LegacyStreamCore.sol";
 import "../smart-contracts/AuctionContract.sol";
 import "./helpers/Assertions.sol";
 import "./helpers/DropAuthTestHelper.sol";